[Breaking] Change null handling for parsing.
Previously, routes would return an empty dictionary if the user had no
data for that gamemode (quickplay or competitive). Now, a simple `null`
is returned instead.

Additionally, the non-relevant fields in routes no longer have summary
dicts, and instead have been replaced with Nones. This will break code
that depends on these routes having the fields, but there was no reason
to use them when not calling `/blob` or the right endpoint.

Signed-off-by: Laura <l@veriny.tf>